refactor: move `handle_future_message()` to `ConversationGuard` [WPB-15710]
